Move `subnets.py` to `bittensor/utils` as it is considered more utility than deprecated.
