+++ bittensor/utils/subnets.py
